dev mode field for client
