+++ packages/core/src/MatchMaker.ts
@@ -1,7 +1,15 @@
 import { ErrorCode } from './Protocol';
 
 import { requestFromIPC, subscribeIPC } from './IPC';
-import { generateId, merge, REMOTE_ROOM_SHORT_TIMEOUT, retry } from './Utils';
+import {
+  cacheRoomHistory,
+  generateId, getRoomCountKey,
+  getRoomHistoryListKey,
+  merge,
+  reloadFromCache,
+  REMOTE_ROOM_SHORT_TIMEOUT,
+  retry
+} from './Utils';
 
 import { RegisteredHandler } from './matchmaker/RegisteredHandler';
 import { Room, RoomInternalState } from './Room';
@@ -53,30 +61,9 @@ export async function setup(_presence?: Presence, _driver?: MatchMakerDriver, _p
   });
 
   presence.hset(getRoomCountKey(), processId, '0');
+
   if(DEV_MODE) {
-    const roomHistoryList = await presence.hgetall(getRoomHistoryListKey());
-    if(roomHistoryList) {
-      for(const [key, value] of Object.entries(roomHistoryList)) {
-        const roomHistory = JSON.parse(value);
-        const tempRoomListingData = await createRoom(roomHistory.roomName, roomHistory.clientOptions);
-        const tempRoom = rooms[tempRoomListingData.roomId];
-        const tempRoomCache = JSON.parse(await presence.hget(getRoomCacheKey(), key));
-
-        // Delete temporary room references from cache
-        delete rooms[tempRoomListingData.roomId];
-        await presence.hdel(getRoomCacheKey(), key);
-        await presence.hdel(getRoomCacheKey(), tempRoomListingData.roomId);
-        await presence.hdel(getRoomHistoryListKey(), tempRoomListingData.roomId);
-
-        // Restore previous roomId and state
-        tempRoom.roomId = key;
-        tempRoom.state = roomHistory.state;
-        tempRoomCache.processId = tempRoomListingData.processId;
-        tempRoomCache.roomId = key;
-        await presence.hset(getRoomCacheKey(), key, JSON.stringify(tempRoomCache));
-        rooms[key] = tempRoom;
-      }
-    }
+    await reloadFromCache(rooms);
   }
 }
 
@@ -283,6 +270,7 @@ export async function createRoom(roomName: string, clientOptions: ClientOptions)
       "clientOptions": clientOptions,
       "roomName": roomName
     }));
+    room.devMode = true;
   }
   return room;
 }
@@ -383,15 +371,7 @@ export async function gracefullyShutdown(): Promise<any> {
 
     return Promise.all(disconnectAll());
   } else {
-    for(const room of Object.values(rooms)) {
-      const roomHistoryResult = await presence.hget(getRoomHistoryListKey(), room.roomId);
-      if(roomHistoryResult) {
-        const roomHistory = JSON.parse(roomHistoryResult);
-        roomHistory["state"] = room.state;
-        await presence.hdel(getRoomHistoryListKey(), room.roomId);
-        await presence.hset(getRoomHistoryListKey(), room.roomId, JSON.stringify(roomHistory));
-      }
-    }
+    await cacheRoomHistory(rooms);
   }
 }
 
@@ -561,14 +541,3 @@ function getProcessChannel(id: string = processId) {
   return `p:${id}`;
 }
 
-function getRoomCountKey() {
-  return 'roomcount';
-}
-
-function getRoomCacheKey() {
-  return 'roomcaches';
-}
-
-function getRoomHistoryListKey() {
-  return 'roomhistory';
-}
