+++ packages/core/src/Utils.ts
@@ -1,6 +1,8 @@
 import nanoid from 'nanoid';
 
 import { debugAndPrintError } from './Debug';
+import { createRoom, presence } from "./MatchMaker";
+import { Room } from "./Room";
 
 // remote room call timeouts
 export const REMOTE_ROOM_SHORT_TIMEOUT = Number(process.env.COLYSEUS_PRESENCE_SHORT_TIMEOUT || 2000);
@@ -106,4 +108,56 @@ export function merge(a: any, ...objs: any[]): any {
     }
   }
   return a;
+}
+
+export async function reloadFromCache(rooms: {[roomId: string]: Room}) {
+  const roomHistoryList = await presence.hgetall(getRoomHistoryListKey());
+  if(roomHistoryList) {
+    for(const [key, value] of Object.entries(roomHistoryList)) {
+      const roomHistory = JSON.parse(value);
+      const tempRoomListingData = await createRoom(roomHistory.roomName, roomHistory.clientOptions);
+      const tempRoom = rooms[tempRoomListingData.roomId];
+      const tempRoomCache = JSON.parse(await presence.hget(getRoomCacheKey(), key));
+
+      // Delete temporary room references from cache
+      delete rooms[tempRoomListingData.roomId];
+      await presence.hdel(getRoomCacheKey(), key);
+      await presence.hdel(getRoomCacheKey(), tempRoomListingData.roomId);
+      await presence.hdel(getRoomHistoryListKey(), tempRoomListingData.roomId);
+
+      // Restore previous roomId and state
+      tempRoom.roomId = key;
+      tempRoom.state = roomHistory.state;
+      tempRoomCache.processId = tempRoomListingData.processId;
+      tempRoomCache.roomId = key;
+      await presence.hset(getRoomCacheKey(), key, JSON.stringify(tempRoomCache));
+      rooms[key] = tempRoom;
+    }
+  }
+}
+
+export async function cacheRoomHistory(rooms: {[roomId: string]: Room}) {
+  for(const room of Object.values(rooms)) {
+    const roomHistoryResult = await presence.hget(getRoomHistoryListKey(), room.roomId);
+    if(roomHistoryResult) {
+      const roomHistory = JSON.parse(roomHistoryResult);
+      roomHistory["state"] = room.state;
+      roomHistory["clients"] = room.clients;
+      // Rewrite updated room history
+      await presence.hdel(getRoomHistoryListKey(), room.roomId);
+      await presence.hset(getRoomHistoryListKey(), room.roomId, JSON.stringify(roomHistory));
+    }
+  }
+}
+
+export function getRoomCountKey() {
+  return 'roomcount';
+}
+
+export function getRoomCacheKey() {
+  return 'roomcaches';
+}
+
+export function getRoomHistoryListKey() {
+  return 'roomhistory';
 }
